fix: migration

+++ migrations/init.sql
@@ -3,7 +3,7 @@ CREATE TABLE bonds (
   pubkey TEXT NOT NULL,
   vote_account TEXT NOT NULL,
   authority TEXT NOT NULL,
-  revenue_share INTEGER NOT NULL,
+  revenue_share NUMERIC NOT NULL,
   epoch INTEGER NOT NULL,
   updated_at TIMESTAMP WITH TIME ZONE NOT NULL,
   PRIMARY KEY(id)
